# Conflicts:
#	i21Database/Scripts/Post-Deployment/SM/DefaultData/1_MasterMenu.sql
#	i21Database/dbo/Functions/fnFAValidateAssetDepreciation.sql
#	i21Database/dbo/Stored Procedures/uspAPAddPaymentSchedules.sql
#	i21Database/dbo/Stored Procedures/uspARPopulateItemsForCosting.sql
#	i21Database/dbo/Stored Procedures/uspARUpdateInvoiceIntegrations.sql
#	i21Database/dbo/Stored Procedures/uspCTReportContractDetail.sql
#	i21Database/dbo/Stored Procedures/uspFADepreciateAssetWrapper.sql
#	i21Database/dbo/Stored Procedures/uspFADepreciateMultipleAsset.sql
#	i21Database/dbo/Stored Procedures/uspFADisposeAsset.sql
#	i21Database/dbo/Stored Procedures/uspFASyncBookDepreciation.sql
#	i21Database/dbo/Stored Procedures/uspPRUpdateEmployeeTimeOffHours.sql
#	i21Database/dbo/Tables/tblFAFixedAsset.sql
#	i21Database/dbo/Tables/tblGRStorageInventoryReceipt.sql
#	i21Database/dbo/Views/vyuAPREceiptClearing.sql
#	i21Database/dbo/Views/vyuFADepreciationMethod.sql
#	i21Database/dbo/Views/vyuGRSettlementTaxDetailsSubReport.sql
#	i21Database/dbo/Views/vyuGRTransferChargesClearing.sql
#	i21Database/dbo/Views/vyuGRTransferClearing.sql
#	i21Database/dbo/Views/vyuPREmployeeTimeOff.sql
#	i21Database/i21Database.sqlproj

+++ i21Database/dbo/Stored Procedures/uspIPProcessERPBlendDemand.sql
@@ -19,7 +19,7 @@ BEGIN TRY
 		,@strCreatedBy NVARCHAR(50)
 		,@intCompanyLocationSubLocationId INT
 		,@strError NVARCHAR(MAX)
-		,@intTrxSequenceNo INT
+		,@intTrxSequenceNo BIGINT
 		,@strItemNo NVARCHAR(50)
 		,@intItemId INT
 		,@intUnitMeasureId INT
